Release canditate [MMFiler] Convert Image & Convert Audio jobs [ffDiaporama] Fix sound sync issue for NTFS rendering
git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@268 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/src/ffDiaporama/DlgImageCorrection.cpp
@@ -45,11 +45,7 @@ DlgImageCorrection::DlgImageCorrection(cCompositionObject *TheCurrentTextItem,in
 
     InitialFilteredString=CurrentBrush->Image->BrushFileTransform.FilterToString();
 
-#if defined(Q_OS_WIN32)||defined(Q_OS_WIN64)
     setWindowFlags((windowFlags()|Qt::CustomizeWindowHint|Qt::WindowSystemMenuHint|Qt::WindowMaximizeButtonHint)&(~Qt::WindowMinimizeButtonHint));
-#else
-    setWindowFlags(Qt::Window|Qt::WindowTitleHint|Qt::WindowSystemMenuHint|Qt::WindowMaximizeButtonHint|Qt::WindowMinimizeButtonHint|Qt::WindowCloseButtonHint);
-#endif
 
     if (CurrentBrush->Image) {
         CachedImage=CurrentBrush->Image->ImageAt(true,NULL);
@@ -369,8 +365,8 @@ void DlgImageCorrection::accept() {
     // Check if cached filtered file exist
     QString CachedFile=CurrentBrush->Image->FileName;
     CachedFile=CachedFile.replace("."+QFileInfo(CachedFile).suffix(),"_ffd.jpg");
-    if ((InitialFilteredString!=CurrentBrush->Image->BrushFileTransform.FilterToString())||
-        ((GlobalMainWindow->ApplicationConfig->AllowCachedTransfoImages)&&(!QFileInfo(CachedFile).exists()))) {
+    if ((CurrentBrush->Image->BrushFileTransform.FilterToString()!="")&&((InitialFilteredString!=CurrentBrush->Image->BrushFileTransform.FilterToString())||
+            ((GlobalMainWindow->ApplicationConfig->AllowCachedTransfoImages)&&(!QFileInfo(CachedFile).exists())))) {
         if (QFileInfo(CachedFile).exists()) QFile::remove(CachedFile);
         if (GlobalMainWindow->ApplicationConfig->AllowCachedTransfoImages) {
             QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
